refactor(store): handle store creation in helper
let store focus on managing store
