+++ src/helper/statusioHelper.js
@@ -1,12 +1,14 @@
+import { readable, derived } from 'svelte/store';
+
 const getOverallStatusCodeOf = (statusJson) =>
   statusJson.result.status_overall.status_code;
 
-export const getStatusOf = async (statusId) => {
+const getStatusOf = async (statusId) => {
   const response = await fetch(`https://status.io/1.0/status/${statusId}`);
   return response.json();
 };
 
-export const getStatusTickerDataOf = (rawStatusPair) => {
+const getSummarizedStatusOf = (rawStatusPair) => {
   const { platform, status } = rawStatusPair;
   if (!status) {
     return {
@@ -22,3 +24,28 @@ export const getStatusTickerDataOf = (rawStatusPair) => {
     major: statusCode >= 500,
   };
 };
+
+const storeValueOf = (platform, status = null) => ({ platform, status });
+
+export const createRawStatusStore = (platform, statusId) => {
+  const statusioRawStatusStore = readable(
+    storeValueOf(platform),
+    async (set) => {
+      const updateStatus = async () => {
+        const status = await getStatusOf(statusId);
+        set(storeValueOf(platform, status));
+      };
+
+      await updateStatus();
+      const interval = setInterval(updateStatus, 90_000);
+      return () => clearInterval(interval);
+    }
+  );
+
+  return statusioRawStatusStore;
+};
+
+export const createSummarizedStatusStore = (rawStatusStore) =>
+  derived(rawStatusStore, ($rawStatusPair) =>
+    getSummarizedStatusOf($rawStatusPair)
+  );
