+++ src/store/statusioStore.js
@@ -1,5 +1,7 @@
-import { readable, derived } from 'svelte/store';
-import { getStatusOf, getStatusTickerDataOf } from '../helper/statusioHelper';
+import {
+  createRawStatusStore,
+  createSummarizedStatusStore,
+} from '../helper/statusioHelper';
 
 const generatePair = (platform, statusId) => ({
   platform,
@@ -12,31 +14,9 @@ const platformStatusIdpairs = [
   generatePair("Let's Encrypt", '55957a99e800baa4470002da'),
 ];
 
-const storeValueOf = (platform, status = null) => ({ platform, status });
-
-const createStatusioReadableStore = (platform, statusId) => {
-  const statusioReadableStore = readable(
-    storeValueOf(platform),
-    async (set) => {
-      const updateStatus = async () => {
-        const status = await getStatusOf(statusId);
-        set(storeValueOf(platform, status));
-      };
-
-      await updateStatus();
-      const interval = setInterval(updateStatus, 90_000);
-      return () => clearInterval(interval);
-    }
-  );
-
-  return statusioReadableStore;
-};
-
 export const statusioStores = platformStatusIdpairs.map((pair) =>
-  createStatusioReadableStore(pair.platform, pair.statusId)
+  createRawStatusStore(pair.platform, pair.statusId)
 );
 export const statusioTickerStores = statusioStores.map((rawStatusStore) =>
-  derived(rawStatusStore, ($rawStatusPair) =>
-    getStatusTickerDataOf($rawStatusPair)
-  )
+  createSummarizedStatusStore(rawStatusStore)
 );
